restructuration
New file hierarchy

+++ batch_process_init.js
@@ -15,43 +15,49 @@ async function update_db(client, transactions, marketplace) {
   return 1;
 }
 
-async function process_results(result) {
+async function process_results(result, drop) {
   var nb = result.length;
   var unique_tsion = {};
+  var count = 0, unique = 0;
 
   for (var i = 0; i < nb; i++) {
     // We group transactions by act_digest to avoid duplicate entries
     let hash = result[i].receipt.act_digest;
+    if (hash.toString() in unique_tsion) {
+      unique++;
+    }
     if (typeof result[i].act.data != "string") {
       unique_tsion[hash.toString()] = result[i].act.data;
       unique_tsion[hash.toString()].receiver = result[i].receipt.receiver;
       unique_tsion[hash.toString()].name = result[i].act.name;
       unique_tsion[hash.toString()].account = result[i].act.account;
     } else {
+      count++;
       // raw hex data... this shouldn't happen so if we get here that means the server did something wrong when processing the transaction
     }
   }
 
   // We open the connection to the recording database
-  var url = "mongodb://" + process.env.TRANSAC_USR + ":" + process.env.TRANSAC_PWD + "@" + process.env.TRANSAC_SERVER + ":" + process.env.MONGO_DB_PORT;
+  var url = "mongodb://" + process.env.TRANSAC_USR + ":" + process.env.TRANSAC_PWD + "@" + process.env.TRANSAC_SERVER + ":" + process.env.MONGO_DB_PORT+ "?replicaSet=rs0";
   var options = { useNewUrlParser: true };
   let client;
-    
-  try {
-    // First we drop all the data to avoid duplicate calls
-    client = await MongoClient.connect(url, options);
+  client = await MongoClient.connect(url, options);  
 
-    const db = client.db("transactions");
-    var collections = await db.command( { listCollections: 1 } );
-    var nb_coll = Object.keys(collections.cursor.firstBatch).length;
+  if (drop) {
+    try {
+      // First we drop all the data to avoid duplicate entries
+      const db = client.db("transactions");
+      var collections = await db.command( { listCollections: 1 } );
+      var nb_coll = Object.keys(collections.cursor.firstBatch).length;
 
-    for (var i = 0; i < nb_coll; i++) {
-      var item = collections.cursor.firstBatch[i];
-      const drop = await db.collection(item.name).drop();
+      for (var i = 0; i < nb_coll; i++) {
+        var item = collections.cursor.firstBatch[i];
+        const drop = await db.collection(item.name).drop();
+      }
+
+    } catch (error) {
+      console.log(error);
     }
-
-  } catch (error) {
-    console.log(error);
   }
   // Index of unique Hashes / 1 hash per unique transaction (as 1 transaction can match multiple actions)
   var hashes = Object.keys(unique_tsion);
@@ -66,6 +72,7 @@ async function process_results(result) {
     }
   };
 
+  console.log("excluded: " + count + " duplicate hash: " + unique);
   client.close();
 
   return 1;
@@ -100,7 +107,11 @@ async function run(uri) {
 }
 
 async function main() {
-  console.time("Query time");
+  console.time("Total query time");
+
+  var start_date = "2018-10-02T00:00:00.000Z";
+  var cur_date = new Date(start_date);
+
   dotenv.load();
   let client;
   try {
@@ -109,26 +120,62 @@ async function main() {
     client = await MongoClient.connect(url, options);
     const db = client.db("EOS");
 
+    console.time("Batch 1");
     const read = await db.collection("action_traces")
       .find(
         { $or:
-          [
-            {"receipt.receiver": "newdexpocket"},
-            {"act.account": "newdexpocket"}
-          ]
-        })
-      .limit(10)
+            [
+              {"receipt.receiver": "newdexpocket"},
+              {"act.account": "newdexpocket"}
+            ],
+          "block_time" : {
+            "$lt": cur_date.toISOString()
+          }
+        }
+      )
       .toArray();
 
-    var ok = await process_results(read);
+    var ok = await process_results(read, true);
+    console.timeEnd("Batch 1");
+    console.log("Batch 1: " + cur_date.toISOString() + " #" + read.length);
+
+    var today = new Date(Date.now());
+    var i = 2;
+
+    while (cur_date < today) {
+      let old_date = new Date(cur_date);
+      cur_date.setDate(cur_date.getDate() + 1);
+      console.time("Batch " + i);
+      const read = await db.collection("action_traces")
+        .find(
+          { $or:
+            [
+              {"receipt.receiver": "newdexpocket"},
+              {"act.account": "newdexpocket"}
+            ],
+            "block_time":
+              {
+                "$gte": old_date.toISOString(),
+                "$lt": cur_date.toISOString()
+              }
+          }
+        )
+        .toArray();
+
+      var ok = await process_results(read, false);
+      console.timeEnd("Batch " + i);
+      console.log("Batch " + i++ + ": " + cur_date + " #" + read.length);
+    }
+
     client.close();
-    console.timeEnd("Query time");
-    // Then we plugin a Mongo ChangeStream to forward any new incoming transaction
+    console.timeEnd("Total query time");
+
+    // Then we plug in a Mongo ChangeStream to forward any new incoming transaction
     
-    run(url);
+    //run(url);
 
   } catch (err) {
-    console.log(err.stack);
+    console.log(err);
   }
 }
